chore: Seperate map components
